+++ include/Helper/Constants.hpp
@@ -2,6 +2,6 @@
 #define CONSTANTS_HPP
 
 //Load SIZE_FACTOR from .ini/.lua?
-const float SIZE_FACTOR = 16.f; //determines the size of graphical objects for higher or lower resolution
+const unsigned SIZE_FACTOR = 16; //determines the size of graphical objects for higher or lower resolution
 
 #endif //CONSTANTS_HPP
